+++ freeway-http-jetty/src/test/java/com/jujin/freeway/http/jetty/JettyWebSocketUpgradeTest.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2026 dzb
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.jujin.freeway.http.jetty;
+
+import com.jujin.freeway.commons.coercion.CoercerDefault;
+import com.jujin.freeway.commons.json.JsonCodecDefault;
+import com.jujin.freeway.http.HttpEngine;
+import com.jujin.freeway.http.testkit.WebSocketUpgradeContract;
+
+/** Jetty runs the shared WebSocket upgrade-consultation contract. */
+class JettyWebSocketUpgradeTest extends WebSocketUpgradeContract {
+
+  @Override
+  protected HttpEngine newEngine() {
+    return new JettyWebEngine(new JsonCodecDefault(), new CoercerDefault());
+  }
+
+  @Override
+  protected String engineName() {
+    return "jetty";
+  }
+}